+++ CHANGELOG.md
@@ -10,6 +10,8 @@
   handling of millions with `m`.
 * Enhancements to admin log viewer. Log file metadata (size & last modified) available with
   optional upgrade to `hoist-core >= 13.2`.
+* Added `GridModel.hideGroupedColumns` and `GridModel.showUngroupedColumns` to manage visibility of
+  columns when grouped and ungrouped. Both default to true.
 
 ### 💥 Breaking Changes
 
@@ -32,6 +34,9 @@
   that adds an empty line beneath the range pasted from the clipboard in editable grids.
 * Fixes an issue where NumberInput would initially render blank values when `max` and `min` were set.
 * Fixes an issue where tree maps would always show green for a `heatValue` of zero.
+* Set ag-Grid's `suppressMakeColumnVisibleAfterUnGroup` to `true` to avoid unwanted visibility state
+  changes when `Column.excludeFromChooser` is set to `true`. New `GridModel.hideGroupedColumns` and
+  `GridModel.showUngroupedColumns` will not affect columns that have their visibility or hidden states locked.
 
 ### 📚 Libraries